+++ tests/testthat/test-struct_test.R
@@ -4,7 +4,7 @@ test_that("struct_test should be reproducible (serial version)", {
 
   # Generating data
   set.seed(1231)
-  graph <- rdiffnet(200, 10)
+  graph <- rdiffnet(100, 10)
 
   set.seed(1313)
 
@@ -21,17 +21,17 @@ test_that("struct_test should be reproducible (parallel version)", {
 
   # Generating data
   set.seed(1231)
-  graph <- rdiffnet(200, 10, seed.graph = "small-world")
+  graph <- rdiffnet(150, 5, seed.graph = "small-world")
 
   # In order to make the parallel version reproducible, we need to set the
   # RNG algorithm to be L'Ecuyer so the sequences can be reproduced
 
-  N <- 1e3
-  set.seed(123)
+  N <- 100
+  set.seed(123, "L'Ecuyer-CMRG")
   b1 <- struct_test(graph, function(x) mean(threshold(x), na.rm = TRUE), R=N, ncpus=2,
                  parallel="multicore")
 
-  set.seed(123)
+  set.seed(123, "L'Ecuyer-CMRG")
   b2 <- struct_test(graph, function(x) mean(threshold(x), na.rm = TRUE), R=N, ncpus=2,
                  parallel="multicore")
 
@@ -39,6 +39,6 @@ test_that("struct_test should be reproducible (parallel version)", {
   b1$boot$seed <- 1
   b2$boot$seed <- 1
   b1$graph <- b2$graph <- NULL
-  expect_equal(b1$mean_t, b2$mean_t, tolerance=5e-3)
+  expect_equal(b1$mean_t, b2$mean_t, tolerance=1e-2)
 
 })